Add files via upload
Team Name: T-800
Team Leader : Saksham Goyal 
Email Id: saksham.goyal2018@vitstudent.ac.in

Project Name: Orbius
Members:
1.Saksham Goyal
2.Chaitanya Dev Chauhan : chaitanyadev.chauhan@vitstudent.ac.in
3.Aditi Mondal : aditi.mondal2018@vitstudent.ac.in

 Solution Overview & Description:
Orbius is  a navigation based application which also keeps the users informed about the activities happening in their vicinity and far as well. We could also use it to navigate to a spot using google maps by storing user marker location and later retrieve it for other users.
There is one login per user, also the design of map is custom rather than classic google maps and we don't store the personal location of a user , only the marker.
The map service is provided by google while we have used plug ins like  rxdart and geoflutterfire to make the process of calling more robust. This much reliable as it is faster than the other classic applications.
The purpose of project is to enlighten people about their surroundings so that they could be more well informed over things happening around them and  it could be used to guide the labor to closest relief point and make their journey more comfortable since these locations could be mark on the map.